+++ src/__init__.py
@@ -1,48 +1,3 @@
-"""
-Rosette Identification Toolkit - Main Entry Point
-
-This is a mockup/guideline for the main program structure.
-The actual implementation will be developed in future feature branches.
-"""
-
-import cell_segmentation
-import image_input
-import vertex_detection
-
-def main():
-    """
-    Main entry point for the Rosette Identification Toolkit.
-    
-    This function will orchestrate the entire analysis pipeline:
-    1. Load cell snapshot images
-    2. Segment cells from the image
-    3. Detect vertices and identify rosettes
-    4. Calculate cell metrics (area, perimeter, neighbors, etc.)
-    5. Provide interactive exploration interface
-    6. Export results and visualizations
-    
-    Returns:
-        None
-    """
-    print("Rosette Identification Toolkit")
-    print("=" * 50)   
-    print("Analysis pipeline completed!")
-    print("(This is a mockup - actual implementation coming soon)")
-
-def run_batch_mode(input_path, output_path, config_file=None):
-    """
-    Run analysis in batch mode for processing multiple images.
-    
-    Args:
-        input_path (str): Path to directory containing cell images
-        output_path (str): Path to save analysis results
-        config_file (str, optional): Path to configuration file
-    
-    Returns:
-        None
-    """
-    print(f"Batch Mode - Processing images from: {input_path}")
-    print(f"Results will be saved to: {output_path}")
-
-if __name__ == "__main__":
-    main()
+from . import cell_segmentation
+from . import image_input
+from . import vertex_detection
